requirements.txt
google-genai
fastapi